services/movies_admin/requirements/base.txt
psycopg2-binary==2.9.1
django==3.2.7
django-model-utils==4.1.1
djangorestframework==3.12.4
django-filter==2.4.0